+++ CHANGELOG.md
@@ -6,6 +6,10 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.1.0/).
 
 ## [Unreleased]
 
+### Added
+
+- Test against more openHAB versions
+
 ## [0.3.2] - 2025-04-25
 
 Initial public release